Documentacion, flujo de publicacion y comprobacion de interoperabilidad
- docs/CONFORMIDAD.md: declaracion del apartado 44 con la clase M, las partes
  implementadas, los requisitos delegados en herramientas externas y las seis
  interpretaciones adoptadas cuando el articulado admite mas de una lectura
- docs/ANEXOS-D-F-H.md: cada punto de las tres listas con lo que hace Attacca
  y donde se comprueba
- docs/EXTENSIONES.md: las dos extensiones propias, con el motivo por el que
  ninguna es normativa
- scripts/comprobar-interoperabilidad.sh: acredita el apartado 42 empleando
  solo unzip y sha256sum. Se ejecuta en los tres sistemas en cada cambio
- .github/workflows: comprobacion en las tres plataformas y flujo de
  publicacion con los pasos de firma preparados

Sin avisos de clippy en los tres paquetes.

+++ crates/attacca-core/src/fsx/space.rs
@@ -59,7 +59,10 @@ mod platform {
         }
         // `f_bavail` son los bloques disponibles para un proceso sin privilegios,
         // que es la cifra que corresponde a esta comprobación.
-        Some(stat.f_bavail.saturating_mul(stat.f_frsize as u64))
+        // El tipo de `f_bavail` difiere entre plataformas: u64 en Linux y
+        // u32 en macOS. La conversión es necesaria en una de las dos.
+        #[allow(clippy::unnecessary_cast)]
+        Some((stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64))
     }
 
     // Declaración mínima de `statvfs`. Evita arrastrar la dependencia `libc`
